Simplify const declaration for regex
I'm not sure about this... I noticed this strange literal within this `const` declaration.

If this literal is not needed, please accept the PR.  If it's needed, please reject the PR, and if you have time, explain why the matching that literal within the pattern is needed.  It looks like perhaps a literal for an exception that is possibly no longer needed.